Was meant to do this along time ago. Committed the code with documentation and comments.
