--- Projects/ProbabilisticABM/SimpleABM/ExitGate.py
@@ -1,7 +1,0 @@
-from SimpleABM.Object import Object
-
-
-class ExitGate(Object):
-    def __init__(self, width, height):
-        self.width = width
-        self.height = height
