--- Projects/ProbabilisticABM/SimpleABM/Object.py
@@ -1,5 +1,0 @@
-class Object:
-    def __init__(self, x, y, **kwargs):
-        self.attributes = kwargs
-        self.x = x
-        self.y = y
