--- Projects/ProbabilisticABM/SimpleABM/SimpleAgent.py
@@ -1,14 +1,0 @@
-from SimpleABM.Steppable import Steppable
-
-
-class SimpleAgent(Steppable):
-
-    def __init__(self, **kwargs):
-        self.attributes = kwargs
-        # self.attributes['deactivate'] = self.deactivate
-
-    def deactivate(self):
-        self.attributes['isActive'] = False
-        self.attributes['completed'] = True
-        print('Agent {} has left the station at exit {}'.format(self.attributes['id'],
-                                                                self.attributes['d']))
