--- Projects/ProbabilisticABM/SimpleABM/Steppable.py
@@ -1,7 +1,0 @@
-class Steppable:
-
-    def __init__(self, **kwargs):
-        self.attributes = kwargs
-
-    def get_attributes(self):
-        return self.attributes
